Refactor linear page cache history tracer to have separate tracers for global and cursor local events. Add support for cursor events tracers.
