+++ community/io/src/test/java/org/neo4j/io/pagecache/randomharness/RandomPageCacheTestHarness.java
@@ -44,7 +44,10 @@ import org.neo4j.io.pagecache.PagedFile;
 import org.neo4j.io.pagecache.impl.SingleFilePageSwapperFactory;
 import org.neo4j.io.pagecache.impl.muninn.MuninnPageCache;
 import org.neo4j.io.pagecache.tracing.PageCacheTracer;
+import org.neo4j.io.pagecache.tracing.cursor.DefaultPageCursorTracerSupplier;
 import org.neo4j.io.pagecache.tracing.cursor.PageCursorTracerSupplier;
+import org.neo4j.io.pagecache.tracing.linear.LinearHistoryPageCacheTracerTest;
+import org.neo4j.io.pagecache.tracing.linear.LinearTracers;
 
 /**
  * The RandomPageCacheTestHarness can plan and run random page cache tests, repeatably if necessary, and verify that
@@ -54,7 +57,7 @@ import org.neo4j.io.pagecache.tracing.cursor.PageCursorTracerSupplier;
  * before and after executing the planned test respectively, and it can integrate with the adversarial file system
  * for fault injection, and arbitrary PageCacheTracers.
  *
- * See {@link org.neo4j.test.LinearHistoryPageCacheTracerTest} for an example of how to configure and use the harness.
+ * See {@link LinearHistoryPageCacheTracerTest} for an example of how to configure and use the harness.
  */
 public class RandomPageCacheTestHarness implements Closeable
 {
@@ -92,7 +95,7 @@ public class RandomPageCacheTestHarness implements Closeable
         filePageCount = cachePageCount * 10;
         filePageSize = cachePageSize;
         tracer = PageCacheTracer.NULL;
-        cursorTracerSupplier = PageCursorTracerSupplier.NULL;
+        cursorTracerSupplier = DefaultPageCursorTracerSupplier.INSTANCE;
         commandCount = 1000;
 
         Command[] commands = Command.values();
@@ -148,6 +151,14 @@ public class RandomPageCacheTestHarness implements Closeable
     public void setTracer( PageCacheTracer tracer )
     {
         this.tracer = tracer;
+    }
+
+    /**
+     * Set the page cursor tracers supplier.
+     */
+    public void setCursorTracerSupplier( PageCursorTracerSupplier cursorTracerSupplier )
+    {
+        this.cursorTracerSupplier = cursorTracerSupplier;
     }
 
     /**
@@ -227,7 +238,7 @@ public class RandomPageCacheTestHarness implements Closeable
     /**
      * Set the preparation phase to use. This phase is executed before all the planned commands. It can be used to
      * prepare some file contents, or reset some external state, such as the
-     * {@link org.neo4j.test.LinearHistoryPageCacheTracer}.
+     * {@link LinearTracers}.
      *
      * The preparation phase is executed before each iteration.
      */
